+++ prql-compiler/src/ast/pl/utils.rs
@@ -1,11 +1,11 @@
-use super::{BinOp, Expr, ExprKind};
+use super::{BinOp, BinaryExpr, Expr, ExprKind};
 
 pub fn new_binop(left: Option<Expr>, op: BinOp, right: Option<Expr>) -> Option<Expr> {
     match (left, right) {
         (Some(left), Some(right)) => {
             let left = Box::new(left);
             let right = Box::new(right);
-            Some(Expr::from(ExprKind::Binary { left, op, right }))
+            Some(Expr::from(ExprKind::Binary(BinaryExpr { left, op, right })))
         }
         (left, right) => left.or(right),
     }