+++ prql-compiler/src/sql/dialect.rs
@@ -81,15 +81,17 @@ impl Dialect {
             | Dialect::SQLite
             | Dialect::Postgres
             | Dialect::MySql
-            | Dialect::MsSql => SupportLevel::Supported,
-            Dialect::Generic | Dialect::Ansi | Dialect::BigQuery | Dialect::Snowflake => {
+            | Dialect::Generic
+            | Dialect::ClickHouse => SupportLevel::Supported,
+            Dialect::MsSql | Dialect::Ansi | Dialect::BigQuery | Dialect::Snowflake => {
                 SupportLevel::Unsupported
             }
-            Dialect::Hive | Dialect::ClickHouse => SupportLevel::Nascent,
+            // TODO: remove?
+            Dialect::Hive => SupportLevel::Nascent,
         }
     }
 
-    #[deprecated(note = "Use `Dialect::Variants` instead")]
+    #[deprecated(note = "Use `Dialect::VARIANTS` instead")]
     pub fn names() -> &'static [&'static str] {
         Dialect::VARIANTS
     }
@@ -243,6 +245,11 @@ impl DialectHandler for MySqlDialect {
 impl DialectHandler for ClickHouseDialect {
     fn ident_quote(&self) -> char {
         '`'
+    }
+
+    fn supports_distinct_on(&self) -> bool {
+        // https://clickhouse.com/docs/en/sql-reference/statements/select/distinct
+        true
     }
 }