+++ prql-compiler/src/tests/test_bad_error_messages.rs
@@ -51,5 +51,24 @@ fn test_bad_error_messages() {
        │              ────┬───
        │                  ╰───── Unknown name
     ───╯
-    "###)
+    "###);
+
+    // Really complicated error message for something so fundamental
+    assert_display_snapshot!(compile(r###"
+    select tracks
+    from artists
+    "###).unwrap_err(), @r###"
+    Error:
+       ╭─[:2:5]
+       │
+     2 │ ╭─▶     select tracks
+     3 │ ├─▶     from artists
+       │ │
+       │ ╰────────────────────── main expected type `relation`, but found type `infer -> infer`
+       │
+       │     Help: Have you forgotten an argument to function main?
+       │
+       │     Note: Type `relation` expands to `[tuple_of_scalars]`
+    ───╯
+    "###);
 }