+++ web/book/tests/documentation/main.rs
@@ -7,5 +7,6 @@ mod website;
 use ::prql_compiler::Options;
 
 fn compile(prql: &str) -> Result<String, prql_compiler::ErrorMessages> {
+    anstream::ColorChoice::Never.write_global();
     prql_compiler::compile(prql, &Options::default().no_signature())
 }